+++ src/commands/install.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { npmInstall } from '../utils/npm'
+import { npmInstall } from '../modules/npm'
 
 export async function install() {
   const dependency = await vscode.window.showInputBox({
@@ -14,13 +14,10 @@ export async function install() {
   const YES = 'Yes'
   const NO = 'No'
 
-  const isLibrary = await vscode.window.showQuickPick(
-    [YES, NO],
-    {
-      ignoreFocusOut: true,
-      title: 'Is this a Decentraland library?',
-    }
-  )
+  const isLibrary = await vscode.window.showQuickPick([YES, NO], {
+    ignoreFocusOut: true,
+    title: 'Is this a Decentraland library?',
+  })
 
   return npmInstall(dependency, isLibrary === YES)
 }
