+++ src/modules/bin.spec.ts
