+++ src/modules/npm.spec.ts
